+++ frontend/lost/src/components/pipeline/src/globalComponents/modals/ScriptArgumentsTable.js
@@ -1,5 +1,6 @@
 import React from 'react'
 
+
 const ArgumentsTable = (props) => {
     if(props.data){
         return (